pyproject.toml
[project]
name = "met-office-audit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "duckdb>=1.5.1",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "lab>=8.9",
    "matplotlib>=3.10.8",
    "numpy>=2.4.3",
    "obstore>=0.9.2",
    "rustac>=0.9.8",
    "stactools-met-office-deterministic>=0.4.2",
    "tqdm>=4.67.3",
    "typer>=0.24.1",
]

[project.scripts]
met-office-audit = "met_office_audit.cli:main"

[build-system]
requires = ["uv_build>=0.10.10,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]
